refactor: move all tests under one test file
